Let terminal respond to framebuffer mode change
This patch is the first attempt to use the new framebuffer resize
protocol. The terminal registers a signal handler for mode change
events. On the occurrence of such a signal the old framebuffer gets
released and a new dataspace requested. However, this is where the
support for resizing currently ends. The new terminal size does not get
propagated to the client yet. Also the new terminal size is not yet
considered by the VT implementation.