Linux: Disable RAM quota checks
When using Genode on Linux, the strict handling of RAM quotas as
performed by the framework is not always desired. Because on this
platform, core does not manage physical memory, the resource guarantees
as suggested by quotas are not effective. Also, the quota restrictions
imposed by Genode are not enforced anyway. Any Genode process could just
mmap an arbitrary amount of anonymous memory to exceed the configured
quota. The nuisance of configuring quotas may outweighs the benefit of
having quotas. Hence, this patch removes quota checking.

The test at 'os/run/quotaless' performs certain operations that would
normally not be possible. But with this patch applied, the tests pass.

Note that this patch is not supposed to enter the mainline development
of Genode.