Move server API concept to base framework
This commit introduces the new `Component` interface in the form of the
headers base/component.h and base/entrypoint.h. The os/server.h API
has become merely a compatibilty wrapper and will eventually be removed.
The same holds true for os/signal_rpc_dispatcher.h. The mechanism has
moved to base/signal.h and is now called 'Signal_handler'.

This patch has been tested on base-linux, base-okl4, base-hw, base-nova,
and base-sel4.

Since the patch shuffles headers around, please do a 'make clean' in the
build directory.

Note that this patch changes the default behavior of components. Each
component opens a CAP and SIGNAL session during startup in order to
create the component's 'Entrypoint'. So we have to make sure to adjust
the session routing policies correspondingly. I did this for the
printf.run script but other run scripts may be affected as well.

Issue #1832