base: new header structure for native_capability.h
This patch prepares the unification of the capability types across the
various base platforms. The template that is used on most traditional
(non-capability) kernels has been moved to native_types_tpl.h for now.
Once all platforms use the new type (which will resemble the definition
currently used on base-sel4), we will remove the template.

Issue #1832