window layouter: allow multiple maximized windows
This patch reverts my misguided approach to show a window exclusively
when maximized. This was motivated by the idea to use the layouter as
the basis for a virtual console-like user interface. But it turns out
that the requirements of such an interface clash with the expectations
of windowed user interfaces. The attempt to cover all scenarios with a
single layouter seems futile. Instead, a console interface should better
be implemented by a different layouter.